Rename developer/ directory to apply/
