+++ src/Routes.tsx
@@ -9,10 +9,10 @@ import { getEnvFlags } from './apiDefs/env';
 import { getApiCategoryOrder, getApiDefinitions } from './apiDefs/query';
 import { IApiDescription } from './apiDefs/schema';
 import PageContent from './components/PageContent';
+import ApplyForm from './containers/apply/ApplyForm';
+import ApplySuccess from './containers/apply/ApplySuccess';
 import BetaPage from './containers/Beta';
 import BetaSuccess from './containers/BetaSuccess';
-import ApplyForm from './containers/developer/ApplyForm';
-import ApplySuccess from './containers/developer/ApplySuccess';
 import DisabledApplyForm from './containers/DisabledApplyForm';
 import DocumentationRoot from './containers/documentation/DocumentationRoot';
 import ProviderIntegrationGuide from './containers/documentation/ProviderIntegrationGuide';
